requirements.txt
requests  
python-dotenv
fastapi
uvicorn

langchain 
langchain-core
langchain-chroma
langchain-huggingface
langchain-aws

PyPDF2

awscli  
boto3